Prevent lost interrupts; deal with floating point cast issues
The way that the code was dealing with interrupts, there was a window
in which interrupts would be lost.  This change fixes that window by
never resetting the hardware counter, and instead just allowing it
to run continuously and then just sampling the 'difference' from
the last time it was sampled.

In order to do the necessary "wraparound" processing for the hardware
counter, it is critical that the compiler knows the exact C data type
of the counter so that unsigned subtraction is done using the
appropriate number of bits.  For this reason, I define a new type
for the hardware counter and use that type uniformly every time
the hardware counter is referenced.

While in the code I also noticed that there were a number of conversions
to floating point in which the typecast was either incorrect or,
at best, not conservative enough.  I changed all references (whether
necessary or not) to ensure that we cast to a float before doing
math operations (such as division) that would cause a loss of
accuracy if done as an integer.